Fix bug with number of points in each file; Add restrict command
I found a bug in the way we were iterating through the provenance dict.
Because it wasn't being done in sorted order, the number of values contianed
in one file did not match the number we were getting from the vectors. Now
everything is sorted, when reading, setting the vectors and iterating through
the provenance dict.
I added the 'restrict' command, currently only implement to chosse a subset of
data with error smaller than the given value.