Fix issue in genetic algorithm; First implementation of MultiNest.
MultiNest algorithm is included but for now can only be run from a
system call. There is a new 'nest' command to do this. In this way
MPI is enabled, output is redirected to the OPEN prompt and
hopefully I can implement some sort of live plotting in the future.