Reimplement genetic algorithm.
Decided to switch from galileo.py to the DEAP implementation of a
simple genetic algorithm. This is a more mature library with plenty
of interesting stuff under the hood. Note: this obviously introduces
a dependency on DEAP, but I think is is quite easy to get/install.
There is still a bug as it sometimes halts for no apparent reason.