Improving the periodograms.
Two new options for the per command: --hifac and --fap to set the
lowest frequency (highest period) for the periodogram calculation
and plot the FAP lines, respectively.
Also fixes a bug for periodogram power values smaller than 1e-300
or so, where Matplotlib fails to correctly set the Y limits and
sometimes throws an error. These values are not so uncommon in bls
as strange as they appear.